Debugging issue - ambiguous references

+++ EightQueensRobot/Utilities/AngleUnitConverter.cs
@@ -2,6 +2,6 @@ namespace EightQueensRobot.Utilities;
 
 public static class AngleUnitConverter
 {
-    public static double ToRadians(this float degrees) => degrees * Math.PI / 180;
-    public static double ToDegrees(this float radians) => radians * 180 / Math.PI;
+    public static float ToRadians(this float degrees) => (float)(degrees * Math.PI / 180);
+    public static float ToDegrees(this float radians) => (float)(radians * 180 / Math.PI);
 }
